link errors

+++ chap-01/4-link-errors/3-sub.hpp
@@ -1,3 +1,3 @@
 #pragma once
 
-int sub(int a, int b);
+float sub(float a, float b);
